separate mock_tikv into a server and an in-memory store
