+++ mock-tikv/Cargo.toml
@@ -7,4 +7,5 @@ edition = "2018"
 futures = "0.3"
 grpcio = { version = "0.6", features = [ "secure", "prost-codec" ], default-features = false }
 kvproto = { git = "https://github.com/pingcap/kvproto.git", rev = "1e28226154c374788f38d3a542fc505cd74720f3", features = [ "prost-codec" ], default-features = false }
-
+derive-new = "0.5.8"
+tikv-client-common = { path = "../tikv-client-common"}
